contrib: Add script to test build against all stable trees
Add a script which builds softiwarp's kernel module against the
corresponding kernel tree for every long-term kernel release.  This
script was written for Ubuntu 16.10 and will need to be adjusted for
other Linux distributions.

This is intended to aid in code maintenance by providing an easy way to
verify that the code still builds on all relevant upstream kernel
versions.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>